refactor(chatroom): extract MagenticStrategy prompts and parsers
TS3 + TS7 from the 2026-05-10 codebase review (issue #276). Security-
critical orchestration surface; partial extraction with the runner
methods deferred.

Extractions (all moved verbatim):
- magenticParsers.ts: formatManagerResponse, parseManagerResponse,
  ManagerDecision, failTask
- magenticPrompts.ts: buildPlanPrompt, buildAssignPrompt,
  buildWorkerPrompt, buildSynthesisPrompt as pure functions

BaseStrategy adds requireAbortController() that throws a named error
if the controller is missing rather than producing a deep TypeError
inside the SDK call site. All this.abortController!.signal usages
across MagenticStrategy, GroupChatStrategy, HandoffStrategy, and
SequentialStrategy are replaced (TS7).

Result:
- MagenticStrategy.ts: 744 -> 543 LOC
- 1471 tests passing including all 155 session-group tests
- Zero ! non-null assertions on abortController across all 5 strategies

Out of scope (deferred follow-up):
- runWorkerTask / executeAssignments / resolveAssignments extraction
  to MagenticTaskRunner. These reference this.config, this.workerTimeoutMs,
  this.requireAbortController(), this.currentUnsubs, this.isAborted, plus
  call other private methods. Cleanly extracting them requires explicit
  dependency injection. Will file a follow-up issue.
- TS5 (Zod-validate JSON.parse boundaries — would replace as-cast in
  parseManagerResponse with a proper schema).
- TS9 (substring-based side-effect classifier in approval-gate).

Bumps to v0.53.1 from current master v0.53.0.

Closes #276
